+++ backend/canisters/group/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Support toggle of mute @everyone notifications ([#8496](https://github.com/open-chat-labs/open-chat/pull/8496))
+
 ## [[2.0.1864](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1864-group)] - 2025-08-12
 
 ### Changed